requirements.txt
firebase-admin
fastapi
uvicorn
requests
rembg
pillow
opencv-python
mediapipe
numpy
